pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "marine-domain-rag"
version = "0.1.0"
description = "Marine domain RAG/GraphRAG/LangGraph demo with EXAONE LLM"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "requests>=2.31",
    "tqdm>=4.66",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "scikit-learn>=1.3",
    "sentence-transformers>=2.7",
    "faiss-cpu>=1.7",
    "rank-bm25>=0.2.2",
    "networkx>=3.2",
    "langgraph>=0.0.50",
    "langchain-core>=0.1.30",
    "transformers>=4.40",
    "torch>=2.2",
    "accelerate>=0.27",
    "huggingface_hub>=0.24",
    "llama-cpp-python>=0.3",
]

[project.scripts]
marine-rag = "marine_domain_rag.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
